+++ WebApiSandbox/SandboxAspireServiceDefaults/Extensions.cs
@@ -1,11 +1,9 @@
-using System.Diagnostics;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Diagnostics.HealthChecks;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Diagnostics.HealthChecks;
+using Microsoft.Extensions.Hosting;
 using Microsoft.Extensions.Logging;
-using Microsoft.Extensions.ServiceDiscovery;
-using OpenTelemetry;
 using OpenTelemetry.Metrics;
 using OpenTelemetry.Trace;
 using Serilog;
@@ -14,7 +12,7 @@ using Serilog.Templates.Themes;
 using SerilogTracing;
 using SerilogTracing.Expressions;
 
-namespace Microsoft.Extensions.Hosting;
+namespace SandboxAspireServiceDefaults;
 
 // Adds common .NET Aspire services: service discovery, resilience, health checks, and OpenTelemetry.
 // This project should be referenced by each service project in your solution.
