+++ WebApiSandbox/SandboxRemoteApisImporters/SandboxRemoteApisImporters.csproj
@@ -9,7 +9,7 @@
     <ItemGroup>
       <ProjectReference Include="..\EfCoreContext\EfCoreContext.csproj" />
       <ProjectReference Include="..\SandboxConfigurations\SandboxConfigurations.csproj" />
-      <ProjectReference Include="..\SandboxRemoteApisInterfaces\SandboxRemoteApisInterfaces.csproj" />
+      <ProjectReference Include="..\SandboxRemoteApisImportersInterfaces\SandboxRemoteApisImportersInterfaces.csproj" />
       <ProjectReference Include="..\WebApiSandboxViewModels\WebApiSandboxViewModels.csproj" />
     </ItemGroup>
