+++ WebApiSandbox/WebApiSandbox/Program.cs
@@ -7,11 +7,12 @@ using Microsoft.Extensions.Options;
 using Microsoft.IdentityModel.Tokens;
 using Microsoft.Extensions.Hosting;
 using Microsoft.OpenApi.Models;
+using SandboxAspireServiceDefaults;
 using SandboxAuthentication;
 using SandboxAuthenticationInterfaces;
 using SandboxConfigurations;
-using SandboxRemoteApisInterfaces;
-using SandboxRemoteApisRepositories;
+using SandboxRemoteApisImportersInterfaces;
+using SandboxRemoteApisImporters;
 using SandboxServices;
 using WebApiSandboxControllers;
 using WebApiSandboxRepositories;
