+++ WebApiSandbox/WebApiSandbox/WebApiSandbox.csproj
@@ -19,8 +19,8 @@
         <ProjectReference Include="..\SandboxAspireServiceDefaults\SandboxAspireServiceDefaults.csproj" />
         <ProjectReference Include="..\SandboxAuthentication\SandboxAuthentication.csproj" />
         <ProjectReference Include="..\SandboxConfigurations\SandboxConfigurations.csproj" />
-        <ProjectReference Include="..\SandboxRemoteApisInterfaces\SandboxRemoteApisInterfaces.csproj" />
-        <ProjectReference Include="..\SandboxRemoteApisRepositories\SandboxRemoteApisRepositories.csproj" />
+        <ProjectReference Include="..\SandboxRemoteApisImportersInterfaces\SandboxRemoteApisImportersInterfaces.csproj" />
+        <ProjectReference Include="..\SandboxRemoteApisImporters\SandboxRemoteApisImporters.csproj" />
         <ProjectReference Include="..\SandboxServices\SandboxServices.csproj" />
         <ProjectReference Include="..\WebApiSandboxControllers\WebApiSandboxControllers.csproj" />
         <ProjectReference Include="..\WebApiSandboxRepositories\WebApiSandboxRepositories.csproj" />
